Rsync & Reorganize: Reorder folder names, new gradle deps, add new icons and strings.
