+++ app/src/main/java/com/example/bookworms/adapters/BookAdapter.kt
@@ -5,7 +5,7 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
-import com.example.bookworms.Model.entities.BookEntity
+import com.example.bookworms.models.entities.BookEntity
 import com.example.bookworms.R
 import com.google.android.material.imageview.ShapeableImageView
