+++ build.gradle.kts
@@ -6,7 +6,7 @@ buildscript{
     }
     dependencies {
         classpath("com.google.gms:google-services:4.4.2")
-        classpath("com.android.tools.build:gradle:8.4.2")
+        classpath("com.android.tools.build:gradle:8.5.0")
     }
 }
